requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
filelock==3.8.0
importlib-metadata==4.12.0
Markdown==3.4.1
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
tzdata==2022.4
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
zipp==3.8.1
